thor: Move IrqSlot into x86-specific code
