+++ kernel/thor/arch/arm/thor-internal/arch/system.hpp
@@ -2,8 +2,6 @@
 
 namespace thor {
 
-static inline constexpr int numIrqSlots = 0;
-
 void initializeArchitecture();
 bool isKernelInEl2();
