+++ kernel/thor/arch/x86/thor-internal/arch/system.hpp
@@ -2,8 +2,6 @@
 
 namespace thor {
 
-static inline constexpr int numIrqSlots = 64;
-
 void initializeArchitecture();
 
 } // namespace thor
