+++ kernel/thor/generic/irq.cpp
@@ -11,20 +11,6 @@ namespace {
 }
 
 // --------------------------------------------------------
-// IrqSlot
-// --------------------------------------------------------
-
-void IrqSlot::raise() {
-	assert(_pin);
-	_pin->raise();
-}
-
-void IrqSlot::link(IrqPin *pin) {
-	assert(!_pin);
-	_pin = pin;
-}
-
-// --------------------------------------------------------
 // IrqSink
 // --------------------------------------------------------
